pip_manager.py renamed to python_manager.py
